+++ services/monitor/src/ChainMonitor.ts
@@ -9,11 +9,7 @@ import {
 } from "@ethereum-sourcify/bytecode-utils";
 import type { SourcifyChain } from "@ethereum-sourcify/lib-sourcify";
 import logger from "./logger";
-import type {
-  KnownDecentralizedStorageFetchers,
-  MonitorConfig,
-  SourcifyRequestOptions,
-} from "./types";
+import type { KnownDecentralizedStorageFetchers, MonitorConfig } from "./types";
 import PendingContract from "./PendingContract";
 import type { Logger } from "winston";
 import type SimilarityVerificationClient from "./SimilarityVerificationClient";
@@ -31,7 +27,6 @@ export default class ChainMonitor extends EventEmitter {
   public sourcifyChain: SourcifyChain;
   private sourceFetchers: KnownDecentralizedStorageFetchers;
   private sourcifyServerURLs: string[];
-  private sourcifyRequestOptions: SourcifyRequestOptions;
   private similarityVerificationClient: SimilarityVerificationClient;
 
   private chainLogger: Logger;
@@ -65,7 +60,6 @@ export default class ChainMonitor extends EventEmitter {
     });
 
     this.sourcifyServerURLs = monitorConfig.sourcifyServerURLs;
-    this.sourcifyRequestOptions = monitorConfig.sourcifyRequestOptions;
     this.monitorFactories = monitorConfig.monitorFactories || false;
 
     const chainConfig = {
@@ -398,48 +392,14 @@ export default class ChainMonitor extends EventEmitter {
       });
 
       this.sourcifyServerURLs.forEach(async (url) => {
-        // Setup retry mechanism
-        let sourcifyRequestAttempt = 0;
-        while (
-          sourcifyRequestAttempt < this.sourcifyRequestOptions.maxRetries
-        ) {
-          try {
-            await pendingContract.sendToSourcifyServer(url, creatorTxHash);
-            break;
-          } catch (error: any) {
-            this.chainLogger.error(
-              "Error sending contract to Sourcify server",
-              {
-                url,
-                error,
-                address,
-                sourcifyRequestAttempt,
-              },
-            );
-
-            // If request fails increment number of attempts
-            sourcifyRequestAttempt++;
-
-            if (
-              sourcifyRequestAttempt < this.sourcifyRequestOptions.maxRetries
-            ) {
-              // If number of attempts is less than maxRetries, wait and retry
-              await new Promise((resolve) =>
-                setTimeout(resolve, this.sourcifyRequestOptions.retryDelay),
-              );
-            } else {
-              // If number of attempts >= maxRetries throw Error
-              this.chainLogger.error(
-                "Failed to send contract to Sourcify server after multiple attempts.",
-                {
-                  url,
-                  error,
-                  address,
-                  sourcifyRequestAttempt,
-                },
-              );
-            }
-          }
+        try {
+          await pendingContract.sendToSourcifyServer(url, creatorTxHash);
+        } catch (error: any) {
+          this.chainLogger.error("Error sending contract to Sourcify server", {
+            url,
+            error,
+            address,
+          });
         }
       });
     } catch (error: any) {