+++ services/monitor/test/Monitor.spec.ts
@@ -24,8 +24,6 @@ const HARDHAT_PORT = 8546;
 // Configured in hardhat.config.js
 const HARDHAT_BLOCK_TIME_IN_SEC = 3;
 const MOCK_SOURCIFY_SERVER = "http://mocksourcifyserver.dev/server/";
-const MOCK_SOURCIFY_SERVER_RETURNING_ERRORS =
-  "http://mocksourcifyserver-returning-errors.dev/server/";
 const MOCK_SIMILARITY_SERVER = "http://mocksimilarity.dev/server/";
 const localChain = {
   chainId: 1337,
@@ -241,43 +239,6 @@ describe("Monitor", function () {
         ).to.be.true;
         resolve();
       });
-    });
-  });
-
-  it("should use retry mechanism for failed Sourcify request", (done) => {
-    const maxRetries = 2;
-    monitor = new Monitor([localChain], {
-      sourcifyServerURLs: [MOCK_SOURCIFY_SERVER_RETURNING_ERRORS],
-      sourcifyRequestOptions: {
-        maxRetries,
-        retryDelay: 1000,
-      },
-      chainConfigs: {
-        [localChain.chainId]: {
-          startBlock: 0,
-          blockInterval: HARDHAT_BLOCK_TIME_IN_SEC * 1000,
-        },
-      },
-    });
-
-    deployFromAbiAndBytecode(
-      signer,
-      storageContractArtifact.abi,
-      storageContractArtifact.bytecode,
-      [],
-    ).then(() => {
-      let sourcifyMockTimesCalled = 0;
-      nock(MOCK_SOURCIFY_SERVER_RETURNING_ERRORS)
-        .post("/")
-        .times(maxRetries)
-        .reply(function () {
-          sourcifyMockTimesCalled++;
-          if (sourcifyMockTimesCalled === maxRetries) {
-            done();
-          }
-          return [500];
-        });
-      monitor.start();
     });
   });